expose router engine for other implementations
